+++ lib/clixon/clixon_text_syntax.h
@@ -2,9 +2,7 @@
  *
   ***** BEGIN LICENSE BLOCK *****
  
-  Copyright (C) 2009-2016 Olof Hagsand and Benny Holmgren
-  Copyright (C) 2017-2019 Olof Hagsand
-  Copyright (C) 2020-2022 Olof Hagsand and Rubicon Communications, LLC(Netgate)
+  Copyright (C) 2022 Olof Hagsand and Rubicon Communications, LLC(Netgate)
 
   This file is part of CLIXON.
 
@@ -33,29 +31,17 @@
 
   ***** END LICENSE BLOCK *****
 
- * Clixon XML object vectors
+ * TEXT / curly-brace syntax parsing and translations
  */
-#ifndef _CLIXON_XML_VEC_H
-#define _CLIXON_XML_VEC_H
-
-/*
- * Types
- */
-typedef struct clixon_xml_vec clixon_xvec; /* struct defined in clicon_xml_vec.c */
+#ifndef _CLIXON_TEXT_SYNTAX_H
+#define _CLIXON_TEXT_SYNTAX_H
 
 /*
  * Prototypes
  */
-clixon_xvec *clixon_xvec_new(void);
-clixon_xvec *clixon_xvec_dup(clixon_xvec *xv0);
-int          clixon_xvec_free(clixon_xvec *xv);
-int          clixon_xvec_len(clixon_xvec *xv);
-cxobj       *clixon_xvec_i(clixon_xvec *xv, int i);
-int          clixon_xvec_extract(clixon_xvec *xv, cxobj ***xvcec, int *xlen, int *xmax);
-int          clixon_xvec_append(clixon_xvec *xv, cxobj *x);
-int          clixon_xvec_prepend(clixon_xvec *xv, cxobj *x);
-int          clixon_xvec_insert_pos(clixon_xvec *xv, cxobj *x, int i);
-int          clixon_xvec_rm_pos(clixon_xvec *xv, int i);
-int          clixon_xvec_print(FILE *f, clixon_xvec *xv);
+int xml2txt(cxobj *xn, clicon_output_cb *fn, FILE *f, int level);
+int clixon_text_syntax_parse_string(char *str, yang_bind yb, yang_stmt *yspec, cxobj **xt, cxobj **xerr);
+int clixon_text_syntax_parse_file(FILE *fp, yang_bind yb, yang_stmt *yspec, cxobj **xt, cxobj **xerr);
 
-#endif /* _CLIXON_XML_VEC_H */
+#endif /* _CLIXON_TEXT_SYNTAX_H */
+