+++ lib/clixon/clixon_xml.h
@@ -225,6 +225,7 @@ cxobj    *xml_child_each(cxobj *xparent, cxobj *xprev,  enum cxobj_type type);
 int       xml_child_insert_pos(cxobj *x, cxobj *xc, int i);
 int       xml_childvec_set(cxobj *x, int len);
 cxobj   **xml_childvec_get(cxobj *x);
+int       clixon_child_xvec_append(cxobj *x, clixon_xvec *xv);
 cxobj    *xml_new(char *name, cxobj *xn_parent, enum cxobj_type type);
 cxobj    *xml_new_body(char *name, cxobj *parent, char *val);
 yang_stmt *xml_spec(cxobj *x);