+++ lib/src/clixon_xpath_optimize.c
@@ -342,7 +342,7 @@ xpath_optimize_check(xpath_tree *xs,
     if ((ret = xpath_list_optimize_fn(xs, xv, xvec)) < 0)
 	return -1;
     if (ret == 1){
-	if (clixon_xvec_extract(xvec, xvec0, xlen0) < 0)
+	if (clixon_xvec_extract(xvec, xvec0, xlen0, NULL) < 0)
 	    return -1;
 	clixon_xvec_free(xvec);
 	_optimize_hits++;